requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.5
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.3
MarkupSafe==2.1.1
packaging==21.3
Pillow==8.3.2
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
